Rename: 공용 api 호출 함수 파일 위치 변경
